requirements.txt
pip3 asgiref
autopep8
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
django-ca
cfdiclient
psycopg2-binary
python-dateutil
